refactor: inline getCommitMessage in action.js
The dedicated module was overkill for a 20-line helper used in a single
place. Inlining keeps action.js self-contained without losing readability.
The integration test in action.test.js is unaffected because it asserts on
run() behavior, not on the module boundary.
